fix(delegate): emit steps in the shape the card renders, not bare tool names
`DelegateStep` has been declared since the delegation card was built —
`{kind, text, toolName, toolInput, content, outcome, durationMs}`, with a
comment saying it is "the same shape the spawn_subagent card renders". The
producer never filled it: delegate-api declared `const steps: string[]` and
pushed bare tool names.

The card branches on `kind` to decide how to draw a row. A plain string has no
`kind`, so it fell through to the tool branch with no name, no input and no
result — a column of empty "tool" rows. It stayed hidden because the delegate
card did not render steps at all; they only ever fed a length count.

So: steps are now `DelegateStep` objects carrying the tool name and, when the
SSE consumer measured one, its duration. A peer's `spawn_subagent` appears here
like any other tool call — the sub-agent's own execution lives in the peer's
session, reachable by opening it.

`DelegateStep` also had TWO declarations (core/tool-registry and, implicitly,
`string[]` in shared/agent-delegate). One definition now, in shared, where
`DelegateResponse` lives; tool-registry re-exports it. That is what made the
compiler point at the four wrong call sites the moment the type was corrected.

Tests: 300 files / 6693 cases pass.
